Some minor improvements
- mark the grid when moving objects with arrow keys
- ability to search for objects by name to drag them into the score
- duplicating registry objects (SynthDefs and ProcessDefs)
- make instances from different sub scores selectable at the same time

+++ src/main/kotlin/xenakis/model/player/Recorder.kt
@@ -89,7 +89,7 @@ class Recorder(private val context: Context) {
                 +"s.record(bus: ${bus.superColliderName})"
             }
         }
-        client.send("schedule", listOf(settings.scLangLatency.now, code))
+        client.send("schedule", listOf(settings.scLangLatency.now.toDouble(), code))
     }
 
     private fun pathForNextRecording(): File {
